+++ package/Editor/Tables/LocalizedAssetTableEditor.cs
@@ -46,7 +46,7 @@ namespace UnityEditor.Localization
             if (atf != null)
             {
                 // Force the label to update itself.
-                atf.value = atf.value;
+                atf.RefreshLabels();
             }
         }