+++ package/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
+## [0.2.3] - 2019-02-20
+-Fixed adding Locales and Tables to Addressables not working when the Addressables settings asset did not exist(LOC-24).
+-Fixed exception when changing a table name in the Table Editor window.
+-Fixed AssetTablesField exception when no Asset Tables exist in the project.
+-Updated to Addressables version '0.5.3-preview'.
+
 ## [0.2.2] - 2018-12-17
 -Fixed NullRefException when opening the Asset Tables window and no Asset Tables exist in the project